+++ src/domain/community/user-preferences/index.ts
@@ -1,3 +1,4 @@
 export * from './user.preference.definition.entity';
 export * from './user.preference.definition.interface';
 export * from './user.preference.service';
+export * from './user.preference.module';
