+++ src/domain/community/user-preferences/user.preference.definition.entity.ts
