+++ src/domain/community/user-preferences/user.preference.definition.interface.ts
