+++ src/domain/community/user-preferences/user.preference.entity.ts
