+++ src/domain/community/user-preferences/user.preference.interface.ts
