+++ src/domain/community/user/user.entity.ts
@@ -5,7 +5,7 @@ import { IUser } from '@domain/community/user/user.interface';
 import { Application } from '@domain/community/application/application.entity';
 import { Agent } from '@domain/agent/agent/agent.entity';
 import { NameableEntity } from '@domain/common/entity/nameable-entity';
-import { UserPreference } from '../preferences/user.preference.entity';
+import { UserPreference } from '../user-preferences/user.preference.entity';
 
 @Entity()
 export class User extends NameableEntity implements IUser {
